refactor: fare 패키지 이동
