+++ src/main/java/subway/domain/fare/FareCalculator.java
@@ -1,4 +1,6 @@
-package subway.domain;
+package subway.domain.fare;
+
+import subway.domain.Distance;
 
 public interface FareCalculator {
