+++ src/main/java/subway/domain/fare/strategy/InitialAdditionalFareStrategy.java
@@ -1,12 +1,14 @@
-package subway.domain;
+package subway.domain.fare.strategy;
 
-import static subway.domain.SubwayFareCalculator.ADDITIONAL_FARE;
-import static subway.domain.SubwayFareCalculator.DEFAULT_DISTANCE;
-import static subway.domain.SubwayFareCalculator.DEFAULT_FARE;
-import static subway.domain.SubwayFareCalculator.DISTANCE_THRESHOLD;
-import static subway.domain.SubwayFareCalculator.INITIAL_INCREASE_DISTANCE;
+import static subway.domain.fare.SubwayFareCalculator.ADDITIONAL_FARE;
+import static subway.domain.fare.SubwayFareCalculator.DEFAULT_DISTANCE;
+import static subway.domain.fare.SubwayFareCalculator.DEFAULT_FARE;
+import static subway.domain.fare.SubwayFareCalculator.DISTANCE_THRESHOLD;
+import static subway.domain.fare.SubwayFareCalculator.INITIAL_INCREASE_DISTANCE;
 
 import org.springframework.stereotype.Component;
+import subway.domain.Distance;
+import subway.domain.fare.Fare;
 
 @Component
 public class InitialAdditionalFareStrategy implements FareStrategy {
