+++ src/test/java/subway/domain/fare/FareTest.java
@@ -1,4 +1,4 @@
-package subway.domain;
+package subway.domain.fare;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
